Conflicts:
	Projects/Common/SKDDriver/SKDDatabaseService/DataAccess/SKD.dbml
	Projects/Common/SKDDriver/SKDDatabaseService/DataAccess/SKD.dbml.layout
	Projects/Common/SKDDriver/SKDDatabaseService/DataAccess/SKD.designer.cs
	Projects/Deploy/Scripts/SKUD_Create.sql

+++ Projects/Common/FiresecServiceAPI/INamedBase.cs
@@ -2,7 +2,7 @@
 {
 	public interface INamedBase
 	{
-		int No { get; set; }
+		ushort No { get; set; }
 		string Name { get; set; }
 		string Description { get; set; }
 		string PresentationName { get; }
